% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoneyValue.R
\name{as.character.tinkoff.invest.schemas.MoneyValue}
\alias{as.character.tinkoff.invest.schemas.MoneyValue}
\title{MoneyValue в character}
\usage{
\method{as.character}{tinkoff.invest.schemas.MoneyValue}(x)
}
\arguments{
\item{x}{Объект класса 'tinkoff.invest.schemas.MoneyValue'}
}
\value{
Результат в формате \code{"123.6 rub"}, \code{"233.8 usd"} и т.п.
}
\description{
Переводим класс \href{https://tinkoff.github.io/investAPI/operations/#moneyvalue}{MoneyValue}
(денежная сумма в определенной валюте) в \link[base:character]{base::character}.
}
