% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quotation.R
\name{as.character.tinkoff.invest.schemas.Quotation}
\alias{as.character.tinkoff.invest.schemas.Quotation}
\title{Перевеод в строку объекта Quotation}
\usage{
\method{as.character}{tinkoff.invest.schemas.Quotation}(x)
}
\arguments{
\item{x}{объект класса 'tinkoff.invest.schemas.Quotation'}
}
\description{
Переводим объект класса \href{https://tinkoff.github.io/investAPI/operations/#quotation}{Quotation}
в \code{\link[base:character]{base::character()}}
}
