% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_instruments_currencies.R
\name{as.data.frame.tinkoff.invest.schemas.CurrenciesResponse}
\alias{as.data.frame.tinkoff.invest.schemas.CurrenciesResponse}
\title{Переводим список валют в data.frame}
\usage{
\method{as.data.frame}{tinkoff.invest.schemas.CurrenciesResponse}(x)
}
\arguments{
\item{x}{Объект класса 'tinkoff.invest.schemas.CurrenciesResponse'. Подробнее
смотрите \href{https://tinkoff.github.io/investAPI/instruments/#currenciesresponse}{CurrenciesResponse}}
}
\description{
Переводим список валют в data.frame
}
