% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_users_get_accounts.R
\name{as.data.frame.tinkoff.invest.schemas.GetAccountsResponse}
\alias{as.data.frame.tinkoff.invest.schemas.GetAccountsResponse}
\title{Переводим список аккантов GetAccountsResponse в data.frame}
\usage{
\method{as.data.frame}{tinkoff.invest.schemas.GetAccountsResponse}(x)
}
\description{
Используется pandas для перевода
}
