% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_market_data_get_last_prices.R
\name{as.data.frame.tinkoff.invest.schemas.GetLastPricesResponse}
\alias{as.data.frame.tinkoff.invest.schemas.GetLastPricesResponse}
\title{Переводим GetLastPricesResponse в data.frame}
\usage{
\method{as.data.frame}{tinkoff.invest.schemas.GetLastPricesResponse}(x)
}
\arguments{
\item{x}{Объект класса 'tinkoff.invest.schemas.GetLastPricesResponse'}
}
\description{
Переводим GetLastPricesResponse в data.frame
}
