% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_instruments_get_instrument_by.R
\name{as.data.frame.tinkoff.invest.schemas.InstrumentResponse}
\alias{as.data.frame.tinkoff.invest.schemas.InstrumentResponse}
\title{Переводим InstrumentResponse в data.frame}
\usage{
\method{as.data.frame}{tinkoff.invest.schemas.InstrumentResponse}(x)
}
\arguments{
\item{x}{Объект класса 'tinkoff.invest.schemas.InstrumentResponse'.}
}
\description{
Переводим объект класса \href{https://tinkoff.github.io/investAPI/instruments/#instrumentresponse}{InstrumentResponse}
в data.frame
}
