% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoneyValue.R
\name{as.double.tinkoff.invest.schemas.MoneyValue}
\alias{as.double.tinkoff.invest.schemas.MoneyValue}
\title{Методы для работы с классом tinkoff.invest.schemas.MoneyValue
MoneyValue as.double}
\usage{
\method{as.double}{tinkoff.invest.schemas.MoneyValue}(x)
}
\description{
Методы для работы с классом tinkoff.invest.schemas.MoneyValue
MoneyValue as.double
}
