% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_operations_get_portfolio.R
\name{as.double.tinkoff.invest.schemas.Quotation}
\alias{as.double.tinkoff.invest.schemas.Quotation}
\title{Перевеод в число объекта Quotation}
\usage{
\method{as.double}{tinkoff.invest.schemas.Quotation}(x)
}
\description{
Перевеод в число объекта Quotation
}
