% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoneyValue.R
\name{format.tinkoff.invest.schemas.MoneyValue}
\alias{format.tinkoff.invest.schemas.MoneyValue}
\title{Format MoneyValue}
\usage{
\method{format}{tinkoff.invest.schemas.MoneyValue}(x)
}
\description{
Format MoneyValue
}
