% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quotation.R
\name{print.tinkoff.invest.schemas.Quotation}
\alias{print.tinkoff.invest.schemas.Quotation}
\title{Печать объекта Quotation}
\usage{
\method{print}{tinkoff.invest.schemas.Quotation}(x)
}
\arguments{
\item{x}{объект класса 'tinkoff.invest.schemas.Quotation'}
}
\description{
Выводим на печать объект класса \href{https://tinkoff.github.io/investAPI/operations/#quotation}{Quotation}
}
