% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_accounts.R
\name{rti_accounts}
\alias{rti_accounts}
\title{Загружаем список аккаунтов Тинькофф}
\usage{
rti_accounts(token = Sys.getenv("rti_TOKEN"))
}
\arguments{
\item{token}{ваш токен API Tinkoff Invest}
}
\description{
Загружаем список аккаунтов Тинькофф
}
