% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_instruments_currencies.R
\name{rti_instruments_currencies}
\alias{rti_instruments_currencies}
\title{Список валют}
\usage{
rti_instruments_currencies(token = Sys.getenv("rti_TOKEN"))
}
\arguments{
\item{token}{\href{https://tinkoff.github.io/investAPI/token/}{токен} для работы с TINKOFF INVEST API}
}
\value{
Объект класса 'tinkoff.invest.schemas.CurrenciesResponse'
}
\description{
Получение списка валют по методу
\href{https://tinkoff.github.io/investAPI/instruments/#currencies}{Currencies}
}
