% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_instruments_get_instrument_by.R
\name{rti_instruments_get_instrument_by}
\alias{rti_instruments_get_instrument_by}
\title{Получение данных об инструменте по FIGI}
\usage{
rti_instruments_get_instrument_by(figi, token = Sys.getenv("rti_TOKEN"))
}
\arguments{
\item{figi}{\href{https://en.wikipedia.org/wiki/Financial_Instrument_Global_Identifier}{Financial Instrument Global Identifier}}

\item{token}{\href{https://tinkoff.github.io/investAPI/token/}{токен} для работы с TINKOFF INVEST API}

\item{account_id}{идентификатор счёта пользователя. Список всех счетов можно получить
через вызов \code{\link[=rti_users_get_accounts]{rti_users_get_accounts()}}}
}
\value{
Объект класса 'tinkoff.invest.schemas.InstrumentResponse'.
Подробное описание на \href{https://tinkoff.github.io/investAPI/instruments/#instrumentresponse}{InstrumentResponse}
}
\description{
Используется метод \href{https://tinkoff.github.io/investAPI/instruments/#getinstrumentby}{GetInstrumentBy}
}
