% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_market_data_get_last_prices.R
\name{rti_market_data_get_last_prices}
\alias{rti_market_data_get_last_prices}
\title{Получить последнюю цену по списку FIGI}
\usage{
rti_market_data_get_last_prices(figis, token = Sys.getenv("rti_TOKEN"))
}
\arguments{
\item{figis}{character-ветор со значениями FIGI
\href{https://en.wikipedia.org/wiki/Financial_Instrument_Global_Identifier}{Financial Instrument Global Identifier}}

\item{token}{\href{https://tinkoff.github.io/investAPI/token/}{токен} для работы с TINKOFF INVEST API}
}
\value{
Объект класса 'tinkoff.invest.schemas.GetLastPricesResponse'
См. описание \href{https://tinkoff.github.io/investAPI/marketdata/#getlastpricesresponse}{GetLastPricesResponse}
}
\description{
Получить последнюю цену по списку FIGI
}
