% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_users_get_accounts.R
\name{rti_users_get_accounts}
\alias{rti_users_get_accounts}
\title{Загружаем список аккаунтов Тинькофф}
\usage{
rti_users_get_accounts(token = Sys.getenv("rti_TOKEN"))
}
\arguments{
\item{token}{ваш токен API Tinkoff Invest}
}
\value{
Результат <-  объект класса \code{tinkoff.invest.schemas.GetAccountsResponse}
}
\description{
Используется метод \url{https://tinkoff.github.io/investAPI/users/#getaccountsrequest}
}
