% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_operations_get_portfolio.R
\name{as.data.frame.tinkoff.invest.schemas.PortfolioResponse}
\alias{as.data.frame.tinkoff.invest.schemas.PortfolioResponse}
\title{Перевод объекта PortfolioResponse в data.frame}
\usage{
\method{as.data.frame}{tinkoff.invest.schemas.PortfolioResponse}(x)
}
\arguments{
\item{x}{объект класса 'tinkoff.invest.schemas.PortfolioResponse'}
}
\description{
Конструируем data.frame из данных объекта
\href{https://tinkoff.github.io/investAPI/operations/#portfolioresponse}{PortfolioResponse}
}
