% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_operations_get_portfolio.R
\name{as.double.tinkoff.invest.schemas.MoneyValue}
\alias{as.double.tinkoff.invest.schemas.MoneyValue}
\title{Перевеод в число объекта MoneyValue}
\usage{
\method{as.double}{tinkoff.invest.schemas.MoneyValue}(x)
}
\description{
Перевеод в число объекта MoneyValue
}
