% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoneyValue.R
\name{as.double.tinkoff.invest.schemas.MoneyValue}
\alias{as.double.tinkoff.invest.schemas.MoneyValue}
\title{Перевод MoneyValue в double}
\usage{
\method{as.double}{tinkoff.invest.schemas.MoneyValue}(x)
}
\arguments{
\item{x}{Объект класса 'tinkoff.invest.schemas.MoneyValue'}
}
\description{
Переводим класс \href{https://tinkoff.github.io/investAPI/operations/#moneyvalue}{MoneyValue}
(денежная сумма в определенной валюте) в \link[base:double]{base::double}.
}
