% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoneyValue.R
\name{print.tinkoff.invest.schemas.MoneyValue}
\alias{print.tinkoff.invest.schemas.MoneyValue}
\title{Print MoneyValue}
\usage{
\method{print}{tinkoff.invest.schemas.MoneyValue}(x)
}
\description{
Print MoneyValue
}
