% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rti_operations_get_portfolio.R
\name{rti_operations_get_portfolio}
\alias{rti_operations_get_portfolio}
\title{Получение портфолио}
\usage{
rti_operations_get_portfolio(account_id, token = Sys.getenv("rti_TOKEN"))
}
\value{
Объект класса \code{tinkoff.invest.schemas.PortfolioResponse}
}
\description{
Используется метод \href{https://tinkoff.github.io/investAPI/operations/#portfoliorequest}{GetPortfolio}
}
